Changement de l'id des models de Integer à Long

+++ src/main/java/com/example/gestionnaireepicierie/controllers/payload/request/GroceryDto.java
@@ -6,7 +6,7 @@ import javax.validation.constraints.NotBlank;
 import java.util.List;
 
 public record GroceryDto(
-        @NotBlank int id,
+        @NotBlank Long id,
         @NotBlank String name,
         @NotBlank UserDto owner,
         @NotBlank List<ArticleDto> articles,
